flask/requirements.txt
Click==7.0
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
uWSGI==2.0.18
Werkzeug==0.15.1
Flask-MySQL==1.5.1
PyMySQL==0.10.0
cryptography
mysql-connector
bcrypt
pycryptodome
python-dotenv